+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# next [????.??.??]
+# 1.12.3 [2019.02.09]
 * Support `template-haskell-2.15`.
 * Add a `gshowList` method to `GShow`, which lets us avoid the need for
   `OverlappingInstances` in `Generics.Deriving.TH`. As a consequence, the